pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "nomad-crystallm"
description = "A NOMAD plugin for running CrystaLLM inference in NOMAD installations."
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Ahmed Ilyas", email = "ahmed.ilyas@physik.hu-berlin.de" },
    { name = "Sarthak Kapoor", email = "sarthak.kapoor@physik.hu-berlin.de" },

]
maintainers = [
    { name = "FAIRmat", email = "fairmat@physik.hu-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab[infrastructure]",
    # nomad.orchestrator module depends on infrastructure deps
    "python-magic-bin; sys_platform == 'win32'",
    "nomad-schema-plugin-run>=1.0.1",
    "nomad-schema-plugin-simulation-workflow>=1.0.9",
    "pymatgen>=2023.3.23",
]

[project.urls]
Repository = "https://github.com/FAIRmat-NFDI/nomad-crystallm"

[project.optional-dependencies]
cpu-action = [
    "requests",
    "aiohttp",
    "crystallm @ git+https://github.com/ka-sarthak/CrystaLLM.git@main",
    "robocrys>=0.2.11",
    "transformers>=4.23.1",
    "sentencepiece>=0.1.97",
    "tokenizers>=0.13.1",
    "torchmetrics>=0.11.4",
]
dev = ["ruff", "pytest", "structlog", "python-logstash>=0.4.6",]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLC0415", # `import` should be at the top-level of a file
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
crystallm_inference = "nomad_crystallm.actions:crystallm_inference"
bandgap_prediction = "nomad_crystallm.actions:bandgap_prediction"
crystallm_bandgap_prediction = "nomad_crystallm.actions.crystallm_bandgap_predictor:crystallm_bandgap_prediction"
crystallm_schemas = "nomad_crystallm.schemas:crystallm_schemas"

[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
